refactor(tests): split BalanceDeltaCalculatorTests (455→≤164) (#272, #267, #266)
Split along the existing // MARK: - seams into four @Suite sibling files
(Account, Earmark, Scheduled, EdgeCase) with shared UUIDs/factories in
BalanceDeltaCalculatorTestFixtures. The original file is deleted — every
test migrated to a topical sibling, so there's no reason to keep a slim
remainder.

This removes four baseline entries entirely rather than re-keying them:
- file_length on the 455-line file
- type_body_length on its 392-line struct body
- ~20 identifier_name hits on `tx` (renamed to `txn` throughout)
- 2 force_unwrapping hits on Decimal(string:)!

Fixtures struct is internal because strict_fileprivate disallows fileprivate
and the four sibling suites must share the same UUIDs/instruments.
